+++ lib/pages.ex
@@ -186,6 +186,8 @@ defmodule Pages do
 
   @doc "Visits `path`."
   @spec visit(Pages.Driver.t(), Path.t()) :: Pages.result()
+  @spec visit(Plug.Conn.t(), Path.t()) :: Pages.result()
+  def visit(%Plug.Conn{} = conn, path), do: Pages.new(conn) |> visit(path)
   def visit(%module{} = page, path), do: module.visit(page, path)
 
   @doc """